Format Bayes factors

Format Bayes factors

Usage

format_bf(
  x,
  digits1 = 1,
  digits2 = 2,
  cutoff = NULL,
  italics = TRUE,
  subscript = "10",
  type = "md"
)

Arguments

x

Number for Bayes factor or BayesFactor object

digits1

Number of digits after the decimal for Bayes factors > 1

digits2

Number of digits after the decimal for Bayes factors < 1

cutoff

Cutoff for using _BF_~10~ > <cutoff> or _BF_~10~ < 1 / <cutoff> (value must be > 1)

italics

Logical for whether BF label should be italicized

subscript

Subscript to include with BF label ("10", "01", or "" for no subscript)

type

Type of formatting ("md" = markdown, "latex" = LaTeX)

Value

A character string that includes BF~10~ and then the Bayes factor formatted in Markdown or LaTeX. If Bayes factor is above or below cutoff, _BF_~10~ > <cutoff> or _BF_~10~ < 1 / <cutoff> is used.

Examples

# Format Bayes factors > 1
format_bf(12.4444)
#> [1] "_BF_~10~ = 12.4"
# By default, Bayes factors > 1000 will use scientific notation
format_bf(1244.44)
#> [1] "_BF_~10~ = 1.2×10^3^"
# Control digits for Bayes factors > 1 with digits1
format_bf(1244.44, digits1 = 3)
#> [1] "_BF_~10~ = 1.244×10^3^"
# Control cutoff for scientific notation with cutoff
format_bf(1244.44, cutoff = 10000)
#> [1] "_BF_~10~ = 1244.4"
# Format Bayes factors < 1
format_bf(0.111)
#> [1] "_BF_~10~ = 0.11"
# By default, Bayes factors <_0.001 will use scientific notation
format_bf(0.0001)
#> [1] "_BF_~10~ = 1.0×10^-4^"
# Control digits for Bayes factors < 1 with digits2
format_bf(0.111, digits2 = 3)
#> [1] "_BF_~10~ = 0.111"
# Control cutoff for scientific notation with cutoff
format_bf(0.001, cutoff = 100)
#> [1] "_BF_~10~ < 0.01"